userEdit and routes, sql fixes
