--- views/src/user/components/User.css
@@ -1,12 +1,0 @@
-.user {
-    margin: 1rem;
-    width: calc(45% - 2rem);
-    min-width: 17.5rem;
-}
-
-.user-info h2 {
-    font-size: 1.5rem;
-    margin: 0 0 0.5rem 0;
-    font-weight: normal;
-    color: white;
-}
