--- views/src/user/components/UserParent.css
@@ -1,10 +1,0 @@
-.user-list {
-    list-style: none;
-    margin: 0 auto;
-    padding: 0;
-    width: 50%;
-    max-width: 50rem;
-    display: flex;
-    justify-content: left;
-    flex-wrap: wrap;
-  }
