--- views/src/user/pages/UsersEdit.css
@@ -1,23 +1,0 @@
-.user-edit-info {
-    font-size: 1.5rem;
-    margin: 0 0 0.5rem 0;
-    font-weight: normal;
-    color: white;
-}
-
-/* Temporary solution till the implementation of a proper backdrop for the notification */
-.user-edit__notification-content {
-    margin-top: 15rem;
-    padding: 0;
-    text-align: center;
-}
-  
-.user-edit__notification-actions {
-    text-align: center;
-    margin-bottom: 50rem;
-}
-
-.user-edit__notification-content2 {
-    height: 15rem;
-    width: 100%;
-}
